chore(surveys): split ResolvedSurveyTranslations + add changeset
- Move the ResolvedSurveyTranslations data class to its own file so the
  file-name-matches-class ktlint rule passes naturally; drop the
  @file:Suppress("ktlint:standard:filename") workaround.
- Add a changeset describing the survey translations feature.

Generated-By: PostHog Code
Task-Id: 22f7fe99-61e4-4d54-90c9-24788c083799
